Aufräumen der Klassen (löschen der Dopplungen)
