+++ src/main/java/org/vaadin/example/application/classes/Anleihe.java
@@ -1,4 +1,4 @@
-package org.vaadin.example.application.wertpapier;
+package org.vaadin.example.application.classes;
 
 import java.time.LocalDate;
 import java.util.List;
