+++ src/main/java/org/vaadin/example/application/classes/Ausschuettung.java
@@ -1,26 +1,41 @@
 package org.vaadin.example.application.classes;
 
-import java.sql.Date;
+import java.time.LocalDate;
+/**
+ * Abstrakte Oberklasse für Ausschüttungen wie {@link Dividende} oder {@link Zinszahlung}.
+ *
+ * Eine Ausschüttung stellt eine finanzielle Zahlung dar, die einem Anleger
+ * im Rahmen eines Wertpapiers zufließt. Dazu zählen insbesondere Dividenden (Aktien)
+ * und Zinszahlungen (Anleihen).
+ *
+ * Diese Klasse enthält allgemeine Eigenschaften und Methoden, die allen Ausschüttungsarten gemeinsam sind.
+ *
+ * @author Jan
+ */
 
-public class Ausschuettung {
-    private double ausschuettung;
+public abstract class Ausschuettung{
+    private int ausschüttungId;
     private double betrag;
-    private double steueren;
-    private Date datum;
-    public Ausschuettung(double ausschuettung, double betrag, double steueren, Date datum)
-    {
-        this.ausschuettung=ausschuettung;
-        this.betrag=betrag;
-        this.steueren=steueren;
-        this.datum=datum;
+    private LocalDate datum;
+    private double steuern;
+    private Transaktion transaktion;
+/**
+ * Konstruktor für eine Ausschüttung.
+ */
+ public Ausschuettung(int ausschüttungId, double betrag, LocalDate datum, double steuern, Transaktion transaktion) {
+        this.ausschüttungId = ausschüttungId;
+        this.betrag = betrag;
+        this.datum = datum;
+        this.steuern = steuern;
+        this.transaktion = transaktion;
     }
 
-    public double getAusschuettung() {
-        return ausschuettung;
+    public int getAusschüttungId() {
+        return ausschüttungId;
     }
 
-    public void setAusschuettung(double ausschuettung) {
-        this.ausschuettung = ausschuettung;
+    public void setAusschüttungId(int ausschüttungId) {
+        this.ausschüttungId = ausschüttungId;
     }
 
     public double getBetrag() {
@@ -31,19 +46,27 @@ public class Ausschuettung {
         this.betrag = betrag;
     }
 
-    public double getSteueren() {
-        return steueren;
-    }
-
-    public void setSteueren(double steueren) {
-        this.steueren = steueren;
-    }
-
-    public Date getDatum() {
+    public LocalDate getDatum() {
         return datum;
     }
 
-    public void setDatum(Date datum) {
+    public void setDatum(LocalDate datum) {
         this.datum = datum;
+    }
+
+    public double getSteuern() {
+        return steuern;
+    }
+
+    public void setSteuern(double steuern) {
+        this.steuern = steuern;
+    }
+
+    public Transaktion getTransaktion() {
+        return transaktion;
+    }
+
+    public void setTransaktion(Transaktion transaktion) {
+        this.transaktion = transaktion;
     }
 }
