+++ src/main/java/org/vaadin/example/application/classes/Kauf.java
@@ -1,11 +1,31 @@
 package org.vaadin.example.application.classes;
 
-import java.sql.Date;
+import java.time.LocalDate;
+/**
+ * Repräsentiert einen Kaufvorgang eines {@link Wertpapier} auf einem bestimmten Handelsplatz.
+ *
+ * Diese Klasse erweitert {@link Transaktion} und ergänzt sie um spezifische Informationen
+ * zum Handelsplatz, an dem der Kauf stattgefunden hat (z. B. XETRA, NYSE, etc.).
+ *
+ * Ein Kauf ist eine spezielle Form einer Transaktion, bei der Wertpapiere erworben werden.
+ *
+ * @author Jan
+ */
+public class Kauf extends Transaktion{
+    private String handelsplatz;
+/**
+ * Konstruktor zur Initialisierung eines Kauf-Objekts mit allen Attributen.
+ */
+    public Kauf(String handelsplatz, LocalDate datum, double gebühren, double kurs, int stückzahl, Wertpapier wertpapier, Ausschuettung ausschüttung) {
+        super(datum, gebühren, kurs, stückzahl, wertpapier, ausschüttung);
+        this.handelsplatz = handelsplatz;
+    }
 
-public class Kauf extends Transaktion {
+    public String getHandelsplatz() {
+        return handelsplatz;
+    }
 
-    public Kauf(Date datum, double gebuehren, double kurs, int stueckzahl)
-    {
-        super(datum,gebuehren,kurs,stueckzahl);
+    public void setHandelsplatz(String handelsplatz) {
+        this.handelsplatz = handelsplatz;
     }
 }
