+++ src/main/java/org/vaadin/example/application/classes/Transaktion.java
@@ -1,52 +1,82 @@
 package org.vaadin.example.application.classes;
 
-import java.sql.Date;
+import java.time.LocalDate;
 
-public class Transaktion
-{
-  private Date datum;
-  private double gebuehren;
-  private double kurs;
-  private int stueckzahl;
+/**
+ * Abstrakte Basisklasse für finanzielle Transaktionen eines {@link Wertpapier},
+ * z.B. {@link Kauf} oder {@link Verkauf}.
+ *
+ * Enthält gemeinsame Attribute wie Datum, Kurs, Stückzahl, Gebühren sowie
+ * Referenzen zum zugehörigen Wertpapier und ggf. einer {@link Ausschuettung}.
+ * Diese Klasse dient als Grundlage für konkrete Transaktionstypen.
+ *
+ * @author Jan
+ */
+public abstract class Transaktion {
+    private LocalDate datum;
+    private double gebühren;
+    private double kurs;
+    private int stückzahl;
+    private Wertpapier wertpapier;
+    private Ausschuettung ausschüttung;
 
-  public Date getDatum() {
-    return datum;
-  }
+/**
+ * Konstruktor zur Initialisierung aller Felder einer Transaktion.
+ */
+    public Transaktion(LocalDate datum, double gebühren, double kurs, int stückzahl, Wertpapier wertpapier, Ausschuettung ausschüttung) {
+        this.datum = datum;
+        this.gebühren = gebühren;
+        this.kurs = kurs;
+        this.stückzahl = stückzahl;
+        this.wertpapier = wertpapier;
+        this.ausschüttung = ausschüttung;
+    }
 
-  public void setDatum(Date datum) {
-    this.datum = datum;
-  }
+    public LocalDate getDatum() {
+        return datum;
+    }
 
-  public int getStueckzahl() {
-    return stueckzahl;
-  }
+    public void setDatum(LocalDate datum) {
+        this.datum = datum;
+    }
 
-  public void setStueckzahl(int stueckzahl) {
-    this.stueckzahl = stueckzahl;
-  }
+    public double getGebühren() {
+        return gebühren;
+    }
 
-  public double getKurs() {
-    return kurs;
-  }
+    public void setGebühren(double gebühren) {
+        this.gebühren = gebühren;
+    }
 
-  public void setKurs(double kurs) {
-    this.kurs = kurs;
-  }
+    public double getKurs() {
+        return kurs;
+    }
 
-  public double getGebuehren() {
-    return gebuehren;
-  }
+    public void setKurs(double kurs) {
+        this.kurs = kurs;
+    }
 
-  public void setGebuehren(double gebuehren) {
-    this.gebuehren = gebuehren;
-  }
+    public int getStückzahl() {
+        return stückzahl;
+    }
 
-  public Transaktion(Date datum, double gebuehren, double kurs, int stueckzahl)
-  {
-    this.datum = datum;
-    this.gebuehren = gebuehren;
-    this.kurs = kurs;
-    this.stueckzahl = stueckzahl;
-  }
+    public void setStückzahl(int stückzahl) {
+        this.stückzahl = stückzahl;
+    }
 
+    public Wertpapier getWertpapier() {
+        return wertpapier;
+    }
+
+    public void setWertpapier(Wertpapier wertpapier) {
+        this.wertpapier = wertpapier;
+    }
+
+    public Ausschuettung getAusschüttung() {
+        return ausschüttung;
+    }
+
+    public void setAusschüttung(Ausschuettung ausschüttung) {
+        this.ausschüttung = ausschüttung;
+    }
 }
